prompt method created to get user input
For the human side of the rock, paper, scissors game, a prompt method was used to recieve and store the human input. by using the getElementById prompt, when the user clicked submit, a function called getHumanChoice was called. And whatever they typed within the textbox was then stored in the userMove variable.