commented out old getHumanFunction / created playRound function
I commented out the lines 33 to 36 because I wasn't sure how to fix the scope of that getHumanChoice function, so instead I created a new getHumanChoice function and used a window.prompt() function and embedded it in the variable userMove. From there I created the playRound function that decides a single game of rock paper scissors between the user and the computer.