update to playGame() function
firstly, function playRound was declared inside the playGame() function as well as the variables to keep track of the score of the game between the user and the computer. Lastly, a for loop was added at the end of the function to iterate 5 times. Each iteration getHumanChoice and getComputerChoice was called again via the humanSelection and compSelection variable. Next, the playRound function is called every time so a new round is played each iteration.