restartGame function created
This function was added so that when the score of either player reaches 5, it resets the score of each player, removes the button for each move, and creates a new game button. When that button is clicked the start game button appears again and the user can start the game again.